Auth and hashing done with better security measures

+++ bootiestrappi/src/main/kotlin/com/proj/bootiestrappi/controller/NoteController.kt
@@ -4,6 +4,8 @@ import com.proj.bootiestrappi.Database.model.Note
 import com.proj.bootiestrappi.Database.repo.noteRepo
 import jakarta.websocket.server.PathParam
 import org.bson.types.ObjectId
+import org.springframework.security.core.context.SecurityContext
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -35,25 +37,31 @@ class NoteController (
     )
     @PostMapping
     fun saveNote(@RequestBody body:NoteRequest):NoteResponse{
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val note = repository.save(Note(
             title = body.title,
             color = body.color,
             content = body.content,
             id = body.id?.let { ObjectId(it) }?:ObjectId.get(),
             createdAt = Instant.now(),
-            ownerId = ObjectId()
+            ownerId = ObjectId(ownerId)
         ))
         return note.toNoteResponse()
     }
     @GetMapping
-    fun findByOwnerId(@RequestParam(required = true)ownerId:String):List<NoteResponse>{
+    fun findByOwnerId():List<NoteResponse>{
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return repository.findByOwnerId(ObjectId(ownerId)).map {
             it?.toNoteResponse()!!
         }
     }
     @DeleteMapping("/{id}")
     fun deleteById(@PathVariable id:String){
-        repository.deleteById(ObjectId(id))
+        val note = repository.findById(ObjectId(id)).orElseThrow{ IllegalArgumentException("Note Doesn't exist")}
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        if(note.ownerId.toHexString() == ownerId){
+            repository.deleteById(ObjectId(id))
+        }
     }
 }
